requirements.txt
python-dotenv==0.17.1
requests==2.27.1
pytz==2022.5
WooCommerce==3.0.0
pandas>=2.0.0
black
isort